switch namespace name to use none_or()
Namespace.name was being cast to a string which led to problems for namespace 0
(simply missing from XML dumps for most wikis) which was being set to the
string "None" instead of None.  Switching from `str(name)` to `none_or(name,
str)` fixes the issue.

This is important on wikis that do not include namespace data for each <page>
and where namespace issue must be inferred from the <title> in each page.
Looking titles that start with "None:" is a dead end.